+++ CHANGELOG.md
@@ -6,6 +6,13 @@
     Click to see more.
   </summary>
 
+### Major
+
+- Box [Breaking]: Removed support for `shape` prop values `roundedTop`, `roundedRight`, `roundedBottom`, and `roundedLeft` (#657)
+- Touchable [Breaking]: Removed support for `shape` prop values `roundedTop`, `roundedRight`, `roundedBottom`, and `roundedLeft` (#657)
+- Run codemod for breaking changes: `cd gestalt
+yarn run codemod --parser=flow -t=packages/gestalt-codemods/0.125.0-1.0.0/convert-roundedX.js ~/code/repo`
+
 ### Minor
 
 ### Patch